🔄 [REFACTOR] Remove old component/lib/context files after migration to new structure
