--- frontend/components/error-boundary.tsx
@@ -1,30 +1,0 @@
-'use client'
-import { Component, ReactNode } from 'react'
-
-interface Props { children: ReactNode }
-interface State { hasError: boolean }
-
-export class ErrorBoundary extends Component<Props, State> {
-  state: State = { hasError: false }
-
-  static getDerivedStateFromError(): State {
-    return { hasError: true }
-  }
-
-  render() {
-    if (this.state.hasError) {
-      return (
-        <div className="flex flex-col items-center justify-center min-h-[200px] gap-4">
-          <p className="text-muted-foreground">Service temporarily unavailable — please refresh</p>
-          <button
-            className="px-4 py-2 bg-primary text-primary-foreground rounded"
-            onClick={() => this.setState({ hasError: false })}
-          >
-            Retry
-          </button>
-        </div>
-      )
-    }
-    return this.props.children
-  }
-}
