+++ frontend/components/features/graph/knowledge-graph.tsx
@@ -1,9 +1,9 @@
 'use client'
 import { useEffect, useRef, useState, useMemo } from 'react'
 import { useSession } from 'next-auth/react'
-import { useTopic } from '@/contexts/topic-context'
+import { useTopic } from '@/lib/providers/topic-provider'
 import dynamic from 'next/dynamic'
-import { apiFetch } from '@/lib/api-fetch'
+import { apiFetch } from '@/lib/api/client'
 import { Badge } from '@/components/ui/badge'
 import { Skeleton } from '@/components/ui/skeleton'
 import { Dialog, DialogContent, DialogHeader, DialogTitle } from '@/components/ui/dialog'
