--- frontend/components/session-provider.tsx
@@ -1,7 +1,0 @@
-'use client'
-import { SessionProvider } from 'next-auth/react'
-import { ReactNode } from 'react'
-
-export default function SessionProviderWrapper({ children }: { children: ReactNode }) {
-  return <SessionProvider refetchOnWindowFocus={false}>{children}</SessionProvider>
-}
